[ALPS07829024] MET: smi larb/comm number enhancement
enlarge smi larb/comm number

MTK-Commit-Id: 51673d91d9b354fc4aae63b5034ee033611092fe

CR-Id: ALPS07829024
Feature: MET Driver

+++ common/tinysys/v1/sspm/sspm_met_smi.c
@@ -361,7 +361,7 @@ static int smi_process_argument(const char *__restrict__ arg, int len)
 	uint32_t ipi_buf[3] = {0, 0, 0};
 	uint32_t ret = 0;
 	uint32_t rdata = 0;
-	uint16_t sspm_master = 0;
+	uint32_t sspm_master = 0;
 	uint32_t sspm_meta = 0;
 
 	if (len < 3)
